fix count_elements for series of Composition and improve test coverage in new tests/ptable/test_count_elements.py
split test_ptable.py into plotly and matplotlib submodules
