+++ pymatviz/ptable.py
@@ -88,8 +88,8 @@ def count_elements(
 
     if is_numeric_dtype(srs):
         pass
-    elif is_string_dtype(srs):
-        # Assume all items in values are composition strings
+    elif is_string_dtype(srs) or {*map(type, srs)} <= {str, Composition}:
+        # all items are formula strings or Composition objects
         if count_mode == "occurrence":
             srs = pd.Series(
                 itertools.chain.from_iterable(
