+++ pyproject.toml
@@ -118,7 +118,7 @@ docstring-code-format = true
 
 [tool.ruff.lint.per-file-ignores]
 "__init__.py" = ["F401"]
-"tests/*" = ["D", "FBT001", "FBT002", "PLR2004", "S101"]
+"tests/*" = ["D", "FBT001", "FBT002", "INP001", "PLR2004", "S101"]
 # T201: print found
 # D100: Missing docstring in public module
 "examples/*" = ["D100", "INP001", "T201"]
